Cleanup and add social icons
