+++ apps/docs/components/FooterHelpCallout.tsx
@@ -1,11 +1,13 @@
 const FooterHelpCallout = () => {
   return (
-    <div className="min-w-full px-8 py-6 prose bg-scale-300 dark:bg-whiteA-200 rounded">
-      <h5 className="text-sm m-0 text-scale-1100">Need some help?</h5>
-      <p className="m-0 text-sm text-scale-900">
-        Not to worry, our specialist engineers are here to help. Submit a support ticket through the{' '}
-        <a href="https://app.supabase.com/support/new">Dashboard</a>.
-      </p>
+    <div className="mt-32">
+      <div className="min-w-full px-8 py-6 prose bg-scale-300 dark:bg-whiteA-200 rounded">
+        <h5 className="text-sm m-0 text-scale-1100">Need some help?</h5>
+        <p className="m-0 text-sm text-scale-900">
+          Not to worry, our specialist engineers are here to help. Submit a support ticket through
+          the <a href="https://app.supabase.com/support/new">Dashboard</a>.
+        </p>
+      </div>
     </div>
   )
 }
